requirements.txt
python-dotenv==1.0.0
pdfplumber==0.10.3